+++ src/feat.cpp
@@ -81,6 +81,7 @@ void Feat::readInputFile(const char file[]) {
 			if (tok[0]=="galFile") galFile = tok[1];
 			if (tok[0]=="tileFile") tileFile= tok[1];
 			if (tok[0]=="fibFile") fibFile= tok[1];
+            if (tok[0]=="fibstatusFile") fibstatusFile = tok[1];
             if (tok[0]=="surveyFile") surveyFile= tok[1];
 			if (tok[0]=="outDir") outDir= tok[1];
 			if (tok[0]=="PrintAscii") PrintAscii= s2b(tok[1]);
@@ -91,7 +92,7 @@ void Feat::readInputFile(const char file[]) {
             if (tok[0]=="SStarsfile")SStarsfile=tok[1];
             if (tok[0]=="SkyFfile") SkyFfile=tok[1];
             if (tok[0]=="Secretfile") Secretfile=tok[1];
-
+            if (tok[0]=="runDate") runDate=tok[1];
 
             if (tok[0]=="kind") {
                 Categories = tok.size()-1;