+++ src/feat.h
@@ -30,6 +30,8 @@ class Feat { // F for features
     str SkyFfile;
     str Secretfile;
     str surveyFile;
+    str fibstatusFile;
+    str runDate;
 	bool PrintFits;
     bool PrintAscii;
     bool diagnose;