+++ src/structs.h
@@ -23,17 +23,21 @@ class fpos {
     double fp_y; //y position in mm of positioner
     int location; //location of fiber
     int spectrom;//which spectrometer 0 - 9
+    int stuck=0; // 0=OK
+    int broken=0; // 0=OK
     std::vector<int> N;// Identify neighboring positioners : neighbors of fiber k are N[k] 
     dpair coords; 
 };
 
 class FP  : public std::vector<struct fpos>{ 
     public:
-
     Table fibers_of_sp;// Inv map of spectrom, fibers_of_sp[sp] are fibers of spectrom sp, redundant
 };
 
+    
 FP  read_fiber_positions(const Feat& F);
+void read_fiber_status(FP& FibPos, const Feat& F);
+
 bool int_pairCompare(const std::pair<int, int>& firstElem, const std::pair<int, int>& secondElem);
 
 // galaxy truth -------------------------------------------
@@ -172,6 +176,7 @@ double plate_dist(const double theta);//plate scale conversion
 struct onplate change_coords(const struct target& O, const struct plate& P);
 dpair projection(int g, int j, const MTL& M , const Plates& P); // Projection of g on j
 int num_av_gals(int j, int k, const MTL& M, const Plates& P, const Feat& F, const Assignment& A); // weighted (and only with remaining observation according to the moment in the survey), and doesn't take into account other kinds than QSO LRG ELG not used
+int A_less_than_B(int year_A, int month_A, int day_A, int year_B, int month_B, int day_B);
 
 // Pyplot -----------------------------------------------
 class pyplot {